requirements.txt
numpy
onnxruntime
Pillow
requests
RobotNotice
